Add files via upload
This model was built by using linear regression. In the course of data visualization, it was observed that the marks obtained (in percentage) by a student shared a positive linear relationship with the number of study hours put in. Hence, we chose a linear regression model. The data was split into train and test and finally, an accuracy of 93.69 was obtained with a mean absolute error of roughly 4.18.